project tables created on startup, some refactoring
