+++ database/papers_database_handler.py
@@ -4,6 +4,7 @@ import os
 from dotenv import load_dotenv
 import hashlib
 from utils.status import Status
+from database.database_connection import connect_to_db
 
 load_dotenv()
 
@@ -32,27 +33,6 @@ def _generate_paper_hash(paper_data_dict):
 
     data_string = "||".join(fields_to_hash)
     return hashlib.sha256(data_string.encode('utf-8')).hexdigest()
-
-
-def connect_to_db():
-    """
-    Establishes a connection to the PostgreSQL database.
-    Reads connection parameters from environment variables.
-    """
-    db_host = os.getenv("DB_HOST")
-    db_name = os.getenv("DB_NAME", "papers")
-    db_user = os.getenv("DB_USER")
-    db_password = os.getenv("DB_PASSWORD")
-    db_port = os.getenv("DB_PORT", "5432")
-
-    try:
-        conn = psycopg2.connect(host=db_host, dbname=db_name,
-                                user=db_user, password=db_password, port=db_port)
-        return conn
-    except psycopg2.Error as e:
-        print(f"Error connecting to the database: {e}")
-        return None
-
 
 def insert_papers(papers_data_list):
     """
