+++ llm/tools/Tools_aggregator.py
@@ -1,5 +1,5 @@
 from llm.tools.paper_handling_tools import update_papers, accept, retry_broaden, reformulate_query, detect_out_of_scope_query
-from paper_ranking.paper_ranker import get_best_papers
+from llm.tools.paper_ranker import get_best_papers
 
 tools = [
     update_papers,
